@@ -1,12 +1,16 @@
 module api
 
-go 1.14
+go 1.20
 
 require (
-	github.com/badoux/checkmail v0.0.0-20200623144435-f9f80cb795fa
-	github.com/dgrijalva/jwt-go v3.2.0+incompatible
-	github.com/go-sql-driver/mysql v1.5.0
-	github.com/gorilla/mux v1.7.4
-	github.com/joho/godotenv v1.3.0
-	golang.org/x/crypto v0.0.0-20200709230013-948cd5f35899
+	github.com/go-sql-driver/mysql v1.7.0
+	github.com/gorilla/mux v1.8.0
+	github.com/joho/godotenv v1.5.1
 )
+
+require (
+	github.com/badoux/checkmail v1.2.1
+	golang.org/x/crypto v0.8.0
+)
+
+require github.com/dgrijalva/jwt-go v3.2.0+incompatible
