@@ -4,25 +4,27 @@ import (
 	"api/src/autenticacao"
 	"api/src/banco"
 	"api/src/modelos"
-	"api/src/repositorios"
-	"api/src/respostas"
 	"api/src/seguranca"
-	"encoding/json"
 	"errors"
-	"io/ioutil"
-	"net/http"
 	"strconv"
 	"strings"
+
+	"api/src/repositorios"
+	"api/src/respostas"
+	"encoding/json"
+	"io"
+	"net/http"
 
 	"github.com/gorilla/mux"
 )
 
-// CriarUsuario insere um usuário no banco de dados
-func CriarUsuario(w http.ResponseWriter, r *http.Request) {
-	corpoRequest, erro := ioutil.ReadAll(r.Body)
+//CriarUsuario insere um usuário no banco de dados
+func CriarUsuario(w http.ResponseWriter, r *http.Request){
+	corpoRequest, erro := io.ReadAll(r.Body)
 	if erro != nil {
 		respostas.Erro(w, http.StatusUnprocessableEntity, erro)
 		return
+		
 	}
 
 	var usuario modelos.Usuario
@@ -30,29 +32,31 @@ func CriarUsuario(w http.ResponseWriter, r *http.Request) {
 		respostas.Erro(w, http.StatusBadRequest, erro)
 		return
 	}
-
-	if erro = usuario.Preparar("cadastro"); erro != nil {
+	if erro = usuario.Preparar("cadastro"); erro != nil{
 		respostas.Erro(w, http.StatusBadRequest, erro)
 		return
 	}
+	
 
 	db, erro := banco.Conectar()
 	if erro != nil {
 		respostas.Erro(w, http.StatusInternalServerError, erro)
 		return
 	}
+
 	defer db.Close()
 
 	repositorio := repositorios.NovoRepositorioDeUsuarios(db)
 	usuario.ID, erro = repositorio.Criar(usuario)
 	if erro != nil {
-		respostas.Erro(w, http.StatusInternalServerError, erro)
+		respostas.Erro(w, http.StatusUnprocessableEntity, erro)
 		return
 	}
-
 	respostas.JSON(w, http.StatusCreated, usuario)
-}
 
+	
+
+}
 // BuscarUsuarios busca todos os usuários salvos no banco
 func BuscarUsuarios(w http.ResponseWriter, r *http.Request) {
 	nomeOuNick := strings.ToLower(r.URL.Query().Get("usuario"))
@@ -99,16 +103,14 @@ func BuscarUsuario(w http.ResponseWriter, r *http.Request) {
 
 	respostas.JSON(w, http.StatusOK, usuario)
 }
-
-// AtualizarUsuario altera as informações de um usuário no banco
-func AtualizarUsuario(w http.ResponseWriter, r *http.Request) {
+//AtualiarUsuario altera as informações de um usuário no banco
+func AtualizarUsuario(w http.ResponseWriter, r *http.Request){
 	parametros := mux.Vars(r)
 	usuarioID, erro := strconv.ParseUint(parametros["usuarioId"], 10, 64)
 	if erro != nil {
 		respostas.Erro(w, http.StatusBadRequest, erro)
 		return
 	}
-
 	usuarioIDNoToken, erro := autenticacao.ExtrairUsuarioID(r)
 	if erro != nil {
 		respostas.Erro(w, http.StatusUnauthorized, erro)
@@ -116,11 +118,11 @@ func AtualizarUsuario(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if usuarioID != usuarioIDNoToken {
-		respostas.Erro(w, http.StatusForbidden, errors.New("Não é possível atualizar um usuário que não seja o seu"))
+		respostas.Erro(w, http.StatusForbidden, errors.New("não é possível atualizar um usuário que não seja o seu"))
 		return
 	}
 
-	corpoRequisicao, erro := ioutil.ReadAll(r.Body)
+	corpoRequisicao, erro := io.ReadAll(r.Body)
 	if erro != nil {
 		respostas.Erro(w, http.StatusUnprocessableEntity, erro)
 		return
@@ -128,15 +130,13 @@ func AtualizarUsuario(w http.ResponseWriter, r *http.Request) {
 
 	var usuario modelos.Usuario
 	if erro = json.Unmarshal(corpoRequisicao, &usuario); erro != nil {
-		respostas.Erro(w, http.StatusBadRequest, erro)
-		return
+		respostas.Erro(w, http.StatusBadRequest,erro)
+		return 
 	}
-
 	if erro = usuario.Preparar("edicao"); erro != nil {
-		respostas.Erro(w, http.StatusBadRequest, erro)
+		respostas.Erro(w, http.StatusBadRequest,erro)
 		return
 	}
-
 	db, erro := banco.Conectar()
 	if erro != nil {
 		respostas.Erro(w, http.StatusInternalServerError, erro)
@@ -152,16 +152,14 @@ func AtualizarUsuario(w http.ResponseWriter, r *http.Request) {
 
 	respostas.JSON(w, http.StatusNoContent, nil)
 }
-
-// DeletarUsuario exclui as informações de um usuário no banco
-func DeletarUsuario(w http.ResponseWriter, r *http.Request) {
+//DeletarUsuario exclui as informações de um usuário no banco
+func DeletarUsuario(w http.ResponseWriter, r *http.Request){
 	parametros := mux.Vars(r)
 	usuarioID, erro := strconv.ParseUint(parametros["usuarioId"], 10, 64)
 	if erro != nil {
 		respostas.Erro(w, http.StatusBadRequest, erro)
 		return
 	}
-
 	usuarioIDNoToken, erro := autenticacao.ExtrairUsuarioID(r)
 	if erro != nil {
 		respostas.Erro(w, http.StatusUnauthorized, erro)
@@ -169,7 +167,7 @@ func DeletarUsuario(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if usuarioID != usuarioIDNoToken {
-		respostas.Erro(w, http.StatusForbidden, errors.New("Não é possível deletar um usuário que não seja o seu"))
+		respostas.Erro(w, http.StatusForbidden, errors.New("não é possível deletar um usuário que não seja o seu"))
 		return
 	}
 
@@ -179,16 +177,14 @@ func DeletarUsuario(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	defer db.Close()
-
 	repositorio := repositorios.NovoRepositorioDeUsuarios(db)
 	if erro = repositorio.Deletar(usuarioID); erro != nil {
 		respostas.Erro(w, http.StatusInternalServerError, erro)
 		return
 	}
-
+	
 	respostas.JSON(w, http.StatusNoContent, nil)
 }
-
 // SeguirUsuario permite que um usuário siga outro
 func SeguirUsuario(w http.ResponseWriter, r *http.Request) {
 	seguidorID, erro := autenticacao.ExtrairUsuarioID(r)
@@ -205,7 +201,7 @@ func SeguirUsuario(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if seguidorID == usuarioID {
-		respostas.Erro(w, http.StatusForbidden, errors.New("Não é possível seguir você mesmo"))
+		respostas.Erro(w, http.StatusForbidden, errors.New("não é possível seguir você mesmo"))
 		return
 	}
 
@@ -224,7 +220,6 @@ func SeguirUsuario(w http.ResponseWriter, r *http.Request) {
 
 	respostas.JSON(w, http.StatusNoContent, nil)
 }
-
 // PararDeSeguirUsuario permite que um usuário pare de seguir outro
 func PararDeSeguirUsuario(w http.ResponseWriter, r *http.Request) {
 	seguidorID, erro := autenticacao.ExtrairUsuarioID(r)
@@ -241,7 +236,7 @@ func PararDeSeguirUsuario(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if seguidorID == usuarioID {
-		respostas.Erro(w, http.StatusForbidden, errors.New("Não é possível parar de seguir você mesmo"))
+		respostas.Erro(w, http.StatusForbidden, errors.New("não é possível parar de seguir você mesmo"))
 		return
 	}
 
@@ -286,7 +281,6 @@ func BuscarSeguidores(w http.ResponseWriter, r *http.Request) {
 
 	respostas.JSON(w, http.StatusOK, seguidores)
 }
-
 // BuscarSeguindo traz todos os usuários que um determinado usuário está seguindo
 func BuscarSeguindo(w http.ResponseWriter, r *http.Request) {
 	parametros := mux.Vars(r)
@@ -329,11 +323,11 @@ func AtualizarSenha(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if usuarioIDNoToken != usuarioID {
-		respostas.Erro(w, http.StatusForbidden, errors.New("Não é possível atualizar a senha de um usuário que não seja o seu"))
+		respostas.Erro(w, http.StatusForbidden, errors.New("não é possível atualizar a senha de um usuário que não seja o seu"))
 		return
 	}
 
-	corpoRequisicao, erro := ioutil.ReadAll(r.Body)
+	corpoRequisicao, erro := io.ReadAll(r.Body)
 	var senha modelos.Senha
 	if erro = json.Unmarshal(corpoRequisicao, &senha); erro != nil {
 		respostas.Erro(w, http.StatusBadRequest, erro)
@@ -355,7 +349,7 @@ func AtualizarSenha(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if erro = seguranca.VerificarSenha(senhaSalvaNoBanco, senha.Atual); erro != nil {
-		respostas.Erro(w, http.StatusUnauthorized, errors.New("A senha atual não condiz com a que está salva no banco"))
+		respostas.Erro(w, http.StatusUnauthorized, errors.New("a senha atual não condiz com a que está salva no banco"))
 		return
 	}
 
@@ -372,3 +366,5 @@ func AtualizarSenha(w http.ResponseWriter, r *http.Request) {
 
 	respostas.JSON(w, http.StatusNoContent, nil)
 }
+
+
