@@ -19,7 +19,7 @@ func CriarToken(usuarioID uint64) (string, error) {
 	permissoes["exp"] = time.Now().Add(time.Hour * 6).Unix()
 	permissoes["usuarioId"] = usuarioID
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, permissoes)
-	return token.SignedString([]byte(config.SecretKey))
+	return token.SignedString([]byte("Secret"))
 }
 
 // ValidarToken verifica se o token passado na requisição é valido
@@ -29,16 +29,16 @@ func ValidarToken(r *http.Request) error {
 	if erro != nil {
 		return erro
 	}
-
+	
 	if _, ok := token.Claims.(jwt.MapClaims); ok && token.Valid {
 		return nil
 	}
 
-	return errors.New("Token inválido")
+	return errors.New("token inválido")
 }
 
 // ExtrairUsuarioID retorna o usuarioId que está salvo no token
-func ExtrairUsuarioID(r *http.Request) (uint64, error) {
+func ExtrairUsuarioID(r *http.Request) (uint64, error){
 	tokenString := extrairToken(r)
 	token, erro := jwt.Parse(tokenString, retornarChaveDeVerificacao)
 	if erro != nil {
@@ -54,23 +54,26 @@ func ExtrairUsuarioID(r *http.Request) (uint64, error) {
 		return usuarioID, nil
 	}
 
-	return 0, errors.New("Token inválido")
+	return 0, errors.New("token inválido")
+
 }
 
-func extrairToken(r *http.Request) string {
+func extrairToken(r *http.Request) string{
 	token := r.Header.Get("Authorization")
 
-	if len(strings.Split(token, " ")) == 2 {
-		return strings.Split(token, " ")[1]
+	if len(strings.Split(token, " "))== 2 {
+		return strings.Split(token, " ")[1]  //indice 1, ou seja. segunda palavra
+
 	}
-
 	return ""
-}
 
+}
 func retornarChaveDeVerificacao(token *jwt.Token) (interface{}, error) {
 	if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
-		return nil, fmt.Errorf("Método de assinatura inesperado! %v", token.Header["alg"])
+		return nil, fmt.Errorf("método de assinatura inesperado! %v", token.Header["alg"])
 	}
 
 	return config.SecretKey, nil
 }
+
+
