@@ -10,34 +10,34 @@ import (
 )
 
 var (
-	// StringConexaoBanco é a string de conexão com o MySQL
+	//StringConexaoBanco é a string de conexão com o MySQL
 	StringConexaoBanco = ""
 
-	// Porta onde a API vai estar rodando
+	//Porta onde a API vai estar rodando
 	Porta = 0
 
-	// SecretKey é a chave que vai ser usada para assinar o token
+	//SecretKey é a chave que vai ser usada para assinar o token
 	SecretKey []byte
 )
 
+
 // Carregar vai inicializar as variáveis de ambiente
-func Carregar() {
+func Carregar(){
 	var erro error
 
 	if erro = godotenv.Load(); erro != nil {
 		log.Fatal(erro)
 	}
-
 	Porta, erro = strconv.Atoi(os.Getenv("API_PORT"))
 	if erro != nil {
 		Porta = 9000
 	}
-
 	StringConexaoBanco = fmt.Sprintf("%s:%s@/%s?charset=utf8&parseTime=True&loc=Local",
 		os.Getenv("DB_USUARIO"),
 		os.Getenv("DB_SENHA"),
 		os.Getenv("DB_NOME"),
 	)
-
+	
 	SecretKey = []byte(os.Getenv("SECRET_KEY"))
+
 }
