@@ -9,7 +9,8 @@ import (
 	"github.com/badoux/checkmail"
 )
 
-// Usuario representa um usuário utilizando a rede social
+//Usuario representa um usuário utilizando a rede social
+
 type Usuario struct {
 	ID       uint64    `json:"id,omitempty"`
 	Nome     string    `json:"nome,omitempty"`
@@ -19,44 +20,44 @@ type Usuario struct {
 	CriadoEm time.Time `json:"CriadoEm,omitempty"`
 }
 
-// Preparar vai chamar os métodos para validar e formatar o usuário recebido
+// Preparar vai chamar os métodos para avaliar e formatar o usuário recebido
 func (usuario *Usuario) Preparar(etapa string) error {
 	if erro := usuario.validar(etapa); erro != nil {
 		return erro
 	}
-
 	if erro := usuario.formatar(etapa); erro != nil {
 		return erro
 	}
-
 	return nil
+
 }
 
 func (usuario *Usuario) validar(etapa string) error {
 	if usuario.Nome == "" {
-		return errors.New("O nome é obrigatório e não pode estar em branco")
+		return errors.New("o nome é obrigatório e não pode estar em branco")
 	}
 
 	if usuario.Nick == "" {
-		return errors.New("O nick é obrigatório e não pode estar em branco")
+		return errors.New("o nick é obrigatório e não pode estar em branco")
 	}
-
 	if usuario.Email == "" {
-		return errors.New("O e-mail é obrigatório e não pode estar em branco")
+		return errors.New("o e-mail é obrigatório e não pode estar em branco")
 	}
 
-	if erro := checkmail.ValidateFormat(usuario.Email); erro != nil {
-		return errors.New("O e-mail inserido é inválido")
+	if erro := checkmail.ValidateFormat(usuario.Email); erro != nil{
+	return errors.New("o email inserido é inválido")
+
 	}
 
 	if etapa == "cadastro" && usuario.Senha == "" {
-		return errors.New("A senha é obrigatória e não pode estar em branco")
+		return errors.New("o senha é obrigatório e não pode estar em branco")
 	}
 
 	return nil
+
 }
 
-func (usuario *Usuario) formatar(etapa string) error {
+func (usuario *Usuario) formatar(etapa string) error{
 	usuario.Nome = strings.TrimSpace(usuario.Nome)
 	usuario.Nick = strings.TrimSpace(usuario.Nick)
 	usuario.Email = strings.TrimSpace(usuario.Email)
@@ -66,9 +67,7 @@ func (usuario *Usuario) formatar(etapa string) error {
 		if erro != nil {
 			return erro
 		}
-
 		usuario.Senha = string(senhaComHash)
 	}
-
 	return nil
 }
