@@ -6,20 +6,20 @@ import (
 	"fmt"
 )
 
-// Usuarios representa um repositório de usuarios
-type Usuarios struct {
+//Usuarios representa um repositório de usuarios
+type Usuarios struct{
 	db *sql.DB
 }
 
-// NovoRepositorioDeUsuarios cria um repositório de usuários
-func NovoRepositorioDeUsuarios(db *sql.DB) *Usuarios {
+//NovoRepositorioDeUsuarios cria um repositório de usuários
+func NovoRepositorioDeUsuarios(db *sql.DB) *Usuarios{
 	return &Usuarios{db}
+
 }
 
-// Criar insere um usuário no banco de dados
+//Criar insere um usuário no banco de dados
 func (repositorio Usuarios) Criar(usuario modelos.Usuario) (uint64, error) {
-	statement, erro := repositorio.db.Prepare(
-		"insert into usuarios (nome, nick, email, senha) values(?, ?, ?, ?)",
+	statement, erro := repositorio.db.Prepare("insert into usuarios (nome, nick, email, senha) values(?,?,?,?)",
 	)
 	if erro != nil {
 		return 0, erro
@@ -35,7 +35,6 @@ func (repositorio Usuarios) Criar(usuario modelos.Usuario) (uint64, error) {
 	if erro != nil {
 		return 0, erro
 	}
-
 	return uint64(ultimoIDInserido), nil
 
 }
@@ -103,10 +102,9 @@ func (repositorio Usuarios) BuscarPorID(ID uint64) (modelos.Usuario, error) {
 	return usuario, nil
 }
 
-// Atualizar altera as informações de um usuário no banco de dados
-func (repositorio Usuarios) Atualizar(ID uint64, usuario modelos.Usuario) error {
-	statement, erro := repositorio.db.Prepare(
-		"update usuarios set nome = ?, nick = ?, email = ? where id = ?",
+// Atualizar altera as informações de um usuário no banco
+func (repositorio Usuarios) Atualizar(ID uint64, usuario modelos.Usuario ) error{
+	statement, erro := repositorio.db.Prepare("update usuarios set nome = ?, nick = ?, email = ? where id = ?",
 	)
 	if erro != nil {
 		return erro
@@ -116,12 +114,10 @@ func (repositorio Usuarios) Atualizar(ID uint64, usuario modelos.Usuario) error 
 	if _, erro = statement.Exec(usuario.Nome, usuario.Nick, usuario.Email, ID); erro != nil {
 		return erro
 	}
-
 	return nil
-}
-
-// Deletar exclui as informações de um usuário no banco de dados
-func (repositorio Usuarios) Deletar(ID uint64) error {
+}	
+// Deletar exclui as informações de um usuário no banco
+func (repositorio Usuarios) Deletar(ID uint64) error{
 	statement, erro := repositorio.db.Prepare("delete from usuarios where id = ?")
 	if erro != nil {
 		return erro
@@ -131,12 +127,11 @@ func (repositorio Usuarios) Deletar(ID uint64) error {
 	if _, erro = statement.Exec(ID); erro != nil {
 		return erro
 	}
-
 	return nil
 }
 
-// BuscarPorEmail busca um usuário por email e retorna o seu id e senha com hash
-func (repositorio Usuarios) BuscarPorEmail(email string) (modelos.Usuario, error) {
+//BuscarPorEmail busca um usuário por email e retorna o seu id e senha com hash
+func (repositorio Usuarios) BuscarPorEmail(email string) (modelos.Usuario, error){
 	linha, erro := repositorio.db.Query("select id, senha from usuarios where email = ?", email)
 	if erro != nil {
 		return modelos.Usuario{}, erro
@@ -145,17 +140,15 @@ func (repositorio Usuarios) BuscarPorEmail(email string) (modelos.Usuario, error
 
 	var usuario modelos.Usuario
 
-	if linha.Next() {
+	if linha.Next(){
 		if erro = linha.Scan(&usuario.ID, &usuario.Senha); erro != nil {
 			return modelos.Usuario{}, erro
 		}
+
 	}
-
 	return usuario, nil
-
 }
 
-// Seguir permite que um usuário siga outro
 func (repositorio Usuarios) Seguir(usuarioID, seguidorID uint64) error {
 	statement, erro := repositorio.db.Prepare(
 		"insert ignore into seguidores (usuario_id, seguidor_id) values (?, ?)",
@@ -223,7 +216,6 @@ func (repositorio Usuarios) BuscarSeguidores(usuarioID uint64) ([]modelos.Usuari
 	return usuarios, nil
 
 }
-
 // BuscarSeguindo traz todos os usuários que um determinado usuário está seguindo
 func (repositorio Usuarios) BuscarSeguindo(usuarioID uint64) ([]modelos.Usuario, error) {
 	linhas, erro := repositorio.db.Query(`
@@ -256,7 +248,6 @@ func (repositorio Usuarios) BuscarSeguindo(usuarioID uint64) ([]modelos.Usuario,
 
 	return usuarios, nil
 }
-
 // BuscarSenha traz a senha de um usuário pelo ID
 func (repositorio Usuarios) BuscarSenha(usuarioID uint64) (string, error) {
 	linha, erro := repositorio.db.Query("select senha from usuarios where id = ?", usuarioID)
