@@ -5,25 +5,26 @@ import (
 	"log"
 	"net/http"
 )
-
 // JSON retorna uma resposta em JSON para a requisição
-func JSON(w http.ResponseWriter, statusCode int, dados interface{}) {
+func JSON(w http.ResponseWriter, statusCode int, dados interface{}){
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(statusCode)
 
 	if dados != nil {
-		if erro := json.NewEncoder(w).Encode(dados); erro != nil {
+		if erro := json.NewEncoder(w).Encode(dados); erro != nil{
 			log.Fatal(erro)
 		}
 	}
 
 }
 
-// Erro retorna um erro em formato JSON
+//Erro retorna um erro em formato JSON
 func Erro(w http.ResponseWriter, statusCode int, erro error) {
 	JSON(w, statusCode, struct {
 		Erro string `json:"erro"`
+
 	}{
 		Erro: erro.Error(),
 	})
+
 }
