@@ -15,11 +15,10 @@ type Rota struct {
 	RequerAutenticacao bool
 }
 
-// Configurar coloca todas as rotas dentro do router
-func Configurar(r *mux.Router) *mux.Router {
+//Configurar coloca todas as rotas dentro do router
+func Configurar(r *mux.Router) *mux.Router{
 	rotas := rotasUsuarios
 	rotas = append(rotas, rotaLogin)
-	rotas = append(rotas, rotasPublicacoes...)
 
 	for _, rota := range rotas {
 
@@ -34,4 +33,5 @@ func Configurar(r *mux.Router) *mux.Router {
 	}
 
 	return r
+
 }
