@@ -7,33 +7,33 @@ import (
 
 var rotasUsuarios = []Rota{
 	{
-		URI:                "/usuarios",
-		Metodo:             http.MethodPost,
-		Funcao:             controllers.CriarUsuario,
+		URI:"/usuarios",
+		Metodo: http.MethodPost,
+		Funcao: controllers.CriarUsuario,
 		RequerAutenticacao: false,
 	},
 	{
-		URI:                "/usuarios",
-		Metodo:             http.MethodGet,
-		Funcao:             controllers.BuscarUsuarios,
+		URI:"/usuarios",
+		Metodo: http.MethodGet,
+		Funcao: controllers.BuscarUsuarios,
 		RequerAutenticacao: true,
 	},
 	{
-		URI:                "/usuarios/{usuarioId}",
-		Metodo:             http.MethodGet,
-		Funcao:             controllers.BuscarUsuario,
+		URI:"/usuarios/{usuarioId}",
+		Metodo: http.MethodGet,
+		Funcao: controllers.BuscarUsuario,
 		RequerAutenticacao: true,
 	},
 	{
-		URI:                "/usuarios/{usuarioId}",
-		Metodo:             http.MethodPut,
-		Funcao:             controllers.AtualizarUsuario,
+		URI:"/usuarios/{usuarioId}",
+		Metodo: http.MethodPut,
+		Funcao: controllers.AtualizarUsuario,
 		RequerAutenticacao: true,
 	},
 	{
-		URI:                "/usuarios/{usuarioId}",
-		Metodo:             http.MethodDelete,
-		Funcao:             controllers.DeletarUsuario,
+		URI:"/usuarios/{usuarioId}",
+		Metodo: http.MethodDelete,
+		Funcao: controllers.DeletarUsuario,
 		RequerAutenticacao: true,
 	},
 	{
@@ -66,4 +66,5 @@ var rotasUsuarios = []Rota{
 		Funcao:             controllers.AtualizarSenha,
 		RequerAutenticacao: true,
 	},
+
 }
