@@ -6,8 +6,9 @@ import (
 	"github.com/gorilla/mux"
 )
 
-// Gerar vai retornar um router com as rotas configuradas
-func Gerar() *mux.Router {
+//Gerae vai retornar um router com as rotas configuradas
+func Gerar() *mux.Router{
 	r := mux.NewRouter()
 	return rotas.Configurar(r)
+
 }
